IMP refactor/ Mover carpeta models para arriba. Muchos cambios en rutas de archivos en consecuencia
